# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FinAegis Core Banking API",
        "description": "Open Source Core Banking as a Service - A modern, scalable, and secure core banking platform built with Laravel 12, featuring event sourcing, domain-driven design, workflow orchestration, and quantum-resistant security measures.",
        "contact": {
            "name": "FinAegis Support",
            "email": "support@finaegis.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/exchange-rates/{from}/{to}": {
            "get": {
                "tags": [
                    "Exchange Rates"
                ],
                "summary": "Get current exchange rate",
                "description": "Retrieve the current exchange rate between two assets",
                "operationId": "getExchangeRate",
                "parameters": [
                    {
                        "name": "from",
                        "in": "path",
                        "description": "The source asset code",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "USD"
                        }
                    },
                    {
                        "name": "to",
                        "in": "path",
                        "description": "The target asset code",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "EUR"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "from_asset": {
                                                    "type": "string"
                                                },
                                                "to_asset": {
                                                    "type": "string"
                                                },
                                                "rate": {
                                                    "type": "string"
                                                },
                                                "inverse_rate": {
                                                    "type": "string"
                                                },
                                                "source": {
                                                    "type": "string"
                                                },
                                                "valid_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "is_active": {
                                                    "type": "boolean"
                                                },
                                                "age_minutes": {
                                                    "type": "integer"
                                                },
                                                "metadata": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exchange rate not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/exchange-rates/{from}/{to}/convert": {
            "get": {
                "tags": [
                    "Exchange Rates"
                ],
                "summary": "Convert amount between assets",
                "description": "Convert an amount from one asset to another using current exchange rates",
                "operationId": "convertCurrency",
                "parameters": [
                    {
                        "name": "from",
                        "in": "path",
                        "description": "The source asset code",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "USD"
                        }
                    },
                    {
                        "name": "to",
                        "in": "path",
                        "description": "The target asset code",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "EUR"
                        }
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "The amount to convert (in smallest unit)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10000
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "from_asset": {
                                                    "type": "string"
                                                },
                                                "to_asset": {
                                                    "type": "string"
                                                },
                                                "from_amount": {
                                                    "type": "integer"
                                                },
                                                "to_amount": {
                                                    "type": "integer"
                                                },
                                                "from_formatted": {
                                                    "type": "string"
                                                },
                                                "to_formatted": {
                                                    "type": "string"
                                                },
                                                "rate": {
                                                    "type": "string"
                                                },
                                                "rate_age_minutes": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Conversion not available"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/exchange-rates": {
            "get": {
                "tags": [
                    "Exchange Rates"
                ],
                "summary": "List exchange rates",
                "description": "Get a list of available exchange rates with filtering options",
                "operationId": "listExchangeRates",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Filter by source asset code",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "USD"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Filter by target asset code",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "EUR"
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Filter by rate source",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "manual",
                                "api",
                                "oracle",
                                "market"
                            ]
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Filter by active status",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "valid",
                        "in": "query",
                        "description": "Filter by validity (not expired)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of results per page (max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "from_asset": {
                                                        "type": "string"
                                                    },
                                                    "to_asset": {
                                                        "type": "string"
                                                    },
                                                    "rate": {
                                                        "type": "string"
                                                    },
                                                    "inverse_rate": {
                                                        "type": "string"
                                                    },
                                                    "source": {
                                                        "type": "string"
                                                    },
                                                    "valid_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "expires_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "is_active": {
                                                        "type": "boolean"
                                                    },
                                                    "age_minutes": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "valid": {
                                                    "type": "integer"
                                                },
                                                "stale": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationError": {
                "title": "Validation Error Response",
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "The email field is required."
                            ],
                            "amount": [
                                "The amount must be greater than 0."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "scheme": "bearer"
            }
        }
    },
...
}